Pre-Release Code
We have created a private repository containing an unpolished version of our code as pre-release code, along with a simple guide in README for researchers to use. If you are interested in accessing this repository, please send an email from your affiliated institutional email address to csjk@zju.edu.cn with a brief explanation of your need (e.g., working on a submission for NeurIPS or SIGGRAPH Asia). We will provide you with the link to the repository and access permissions.

Please note that this pre-release is intended for urgent research needs only. If your work is not time-sensitive, we recommend waiting for the code release scheduled for May. The upcoming release will have refinements that are not present in this pre-release, and using the current version might lead to confusion due to mismatches with the improved codebase.